back/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
Pillow==10.1.0
pytesseract==0.3.10
sentencepiece==0.1.99
requests==2.31.0
nltk==3.8.1
opencv-python==4.8.1.78
accelerate==0.27.2